refactor(#217): rename dir with global requests
