+++ src/requests/global/appConfigUpdate.ts
